requirements.txt
alpaca-backtrader-api
black
confluent-kafka
pytest
pytest-dotenv
pytest-xdist
ray
SQLAlchemy==1.4.0b1
websockets
